\name{blanks}
\alias{blanks}
\title{String of Blank Carakters}
\description{
  Create a string of blank characters.
}
\usage{
blanks(n)
}
\arguments{
  \item{n}{integer greater or equal to 0.}
}
\details{
  \code{blanks(n)} is a string of \code{n} blanks.
}
\value{
  String of \code{n} blanks.
}
\seealso{
\code{\link{deblank}}
}
\examples{
blanks(6)
}
\keyword{ string }
