\name{matlab}
\alias{matlab}
\title{
  Matlab Idioms
}
\description{
  Reinstall typical idioms for Matlab functions.
}
\usage{
matlab()
}
\details{
  Reinstalls Matlab names of the following functions:

  \code{find}, \code{fix}, \code{mode}, \code{poly},\cr
  \code{diag}, \code{toeplitz}, \code{reshape}, \code{norm},\cr
  \code{real}, \code{imag},\cr
  \code{strtrim}.
}
\value{
  Invisible NULL value.
}
