\name{matlab}
\alias{matlab}
\title{
  Matlab Idioms
}
\description{
  Reinstall typical idioms for Matlab functions.
}
\usage{
matlab()
}
\details{
  Reinstalls Matlab names of the following functions:

  \code{diag}, \code{find}, \code{fix}, \code{imag}, \code{lcm},\cr
  \code{mode}, \code{norm}, \code{poly}, \code{rank}, \code{real},\cr
  \code{reshape}, \code{strtrim}, \code{toeplitz},

  and generates active bindings for

  \code{ans}, \code{who}, \code{tic}, and \code{toc}.
}
\value{
  Invisible NULL value.
}
